FP2.mk: build TimeKeep
TimeKeep is utility to keep/restore RTC offset for Qualcomm device, where
RTC is not setable. It works by keeping the offset between the RTC time and
system time in a property (or a file) and restore it at boot.

This commit make build system to also build TimeKeep. Coupled with the unit
file added, they'll make sure that time will be correct accross reboots
even when network time isn't available.

Change-Id: Ib62b69bd6a81b30362f32fe80d1557b66675a7a5